+++ packages/menu/src/styles.ts
@@ -1,18 +1,14 @@
-import { css } from '@leafygreen-ui/emotion';
+import { css, cx } from '@leafygreen-ui/emotion';
 import { fontFamilies } from '@leafygreen-ui/tokens';
 import { palette } from '@leafygreen-ui/palette';
+import { Theme } from '@leafygreen-ui/lib';
 import { Size } from './types';
 
 export const svgWidth = 24;
-export const menuItemPadding = 15;
-export const paddingLeft = 52;
-const indentation = 16;
+export const paddingLeftWithGlyph = 54;
+export const paddingLeftWithoutGlyph = 20;
+const indentation = 20;
 const wedgeWidth = 4;
-const minMenuContent = 32;
-const menuBlockPadding: Record<Size, number> = {
-  [Size.Default]: 2,
-  [Size.Large]: 6.5,
-};
 
 /**
  * Base styles
@@ -31,10 +27,9 @@ export const menuItemContainerStyle = css`
   font-size: 13px;
   text-align: left;
   text-decoration: none;
-  color: ${palette.white};
   cursor: pointer;
   border: none;
-  background-color: ${palette.black};
+
   transition: background-color 150ms ease-in-out;
 
   &:focus {
@@ -47,8 +42,6 @@ export const menuItemContainerStyle = css`
     position: absolute;
     width: ${wedgeWidth}px;
     left: 0px;
-    height: calc(100% - 8px);
-    max-height: 32px;
     border-radius: 0 ${wedgeWidth}px ${wedgeWidth}px 0;
     background-color: transparent;
     transition: background-color 150ms ease-in-out;
@@ -56,7 +49,6 @@ export const menuItemContainerStyle = css`
 
   &:hover {
     text-decoration: none;
-    background-color: ${palette.gray.dark3};
   }
 
   &:active {
@@ -64,9 +56,32 @@ export const menuItemContainerStyle = css`
   }
 `;
 
+export const menuItemContainerThemeStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    color: ${palette.white};
+    background-color: ${palette.black};
+
+    &:hover {
+      background-color: ${palette.gray.dark3};
+    }
+  `,
+  [Theme.Dark]: css`
+    color: ${palette.black};
+    background-color: ${palette.gray.light2};
+
+    &:hover {
+      background-color: ${palette.gray.light1};
+    }
+  `,
+};
+
 export const menuItemHeight = (size: Size) => {
   return css`
-    min-height: ${minMenuContent + 2 * menuBlockPadding[size]}px;
+    min-height: ${size === Size.Default ? 36 : 45}px;
+
+    &:before {
+      height: ${size === Size.Default ? 22 : 36}px;
+    }
   `;
 };
 
@@ -78,7 +93,7 @@ export const textContainer = css`
 
 export const mainIconStyle = css`
   color: ${palette.gray.dark1};
-  margin-right: ${paddingLeft - svgWidth - menuItemPadding}px;
+  margin-right: 16px;
   flex-shrink: 0;
 `;
 
@@ -90,7 +105,6 @@ export const titleTextStyle = css`
   width: 100%;
   font-size: 13px;
   font-weight: 500;
-  color: ${palette.white};
 
   // We create a pseudo element that's the width of the bolded text
   // This way there's no layout shift on hover when the text is bolded.
@@ -105,12 +119,27 @@ export const titleTextStyle = css`
   }
 `;
 
-export const descriptionTextStyle = css`
+const descriptionTextStyle = css`
   font-size: 13px;
   font-weight: normal;
   line-height: 16px;
   color: ${palette.gray.light1};
 `;
+
+export const descriptionTextThemeStyle: Record<Theme, string> = {
+  [Theme.Light]: cx(
+    descriptionTextStyle,
+    css`
+      color: ${palette.gray.light1};
+    `,
+  ),
+  [Theme.Dark]: cx(
+    descriptionTextStyle,
+    css`
+      color: ${palette.gray.dark2};
+    `,
+  ),
+};
 
 export const linkDescriptionTextStyle = css`
   white-space: nowrap;
@@ -122,10 +151,17 @@ export const linkDescriptionTextStyle = css`
  * Hover Styles
  */
 
-export const getHoverStyles = (container: string) => ({
+export const getHoverStyles = (container: string, theme: Theme) => ({
   text: css`
-    ${container}:hover & {
+    ${container}:not(:disabled):hover & {
       font-weight: 700;
+    }
+  `,
+  activeText: css`
+    ${container}:not(:disabled):hover & {
+      color: ${theme === Theme.Light
+        ? palette.green.base
+        : palette.green.dark3};
     }
   `,
 });
@@ -133,84 +169,176 @@ export const getHoverStyles = (container: string) => ({
 /**
  * Active styles
  */
-export const activeMenuItemContainerStyle = css`
-  background-color: ${palette.black};
-
-  &:before {
-    background-color: ${palette.green.base};
-  }
-
-  &:hover {
-    color: ${palette.green.base};
+export const activeMenuItemContainerStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    background-color: ${palette.black};
 
     &:before {
       background-color: ${palette.green.base};
     }
-  }
-`;
 
-export const activeTitleTextStyle = css`
-  font-weight: bold;
-  color: ${palette.green.base};
-`;
-export const activeDescriptionTextStyle = css`
-  color: ${palette.gray.light1};
-`;
-export const activeIconStyle = css`
-  color: ${palette.green.base};
-`;
+    &:hover {
+      color: ${palette.green.base};
+
+      &:before {
+        background-color: ${palette.green.base};
+      }
+    }
+  `,
+  [Theme.Dark]: css`
+    background-color: ${palette.gray.light2};
+
+    &:before {
+      background-color: ${palette.green.dark2};
+    }
+
+    &:hover {
+      color: ${palette.green.dark3};
+
+      &:before {
+        background-color: ${palette.green.dark2};
+      }
+    }
+  `,
+};
+
+export const activeTitleTextStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    font-weight: bold;
+    color: ${palette.green.base};
+  `,
+  [Theme.Dark]: css`
+    font-weight: bold;
+    color: ${palette.green.dark2};
+  `,
+};
+
+export const activeDescriptionTextStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    color: ${palette.gray.light1};
+  `,
+  [Theme.Dark]: css`
+    color: ${palette.gray.dark2};
+  `,
+};
+
+export const activeIconStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    color: ${palette.green.base};
+  `,
+  [Theme.Dark]: css`
+    color: ${palette.green.dark2};
+  `,
+};
 
 /**
  * Disabled styles
  */
-export const disabledMenuItemContainerStyle = css`
-  cursor: not-allowed;
-  pointer-events: none;
+export const disabledIconStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    color: ${palette.gray.dark2};
+  `,
+  [Theme.Dark]: css`
+    color: ${palette.gray.light1};
+  `,
+};
 
-  &:hover:before {
-    background-color: unset;
+const disabledMenuItemContainerStyle = css`
+  cursor: not-allowed;
+
+  &:active {
+    pointer-events: none;
+  }
+
+  &:hover {
+    &,
+    &:before {
+      background-color: unset;
+    }
   }
 `;
 
-export const disabledTextStyle = css`
-  color: ${palette.gray.dark1};
-  font-weight: 400;
-`;
+export const disabledMenuItemContainerThemeStyle: Record<Theme, string> = {
+  [Theme.Dark]: cx(
+    disabledMenuItemContainerStyle,
+    css`
+      background-color: ${palette.gray.light2};
+    `,
+  ),
+  [Theme.Light]: cx(
+    disabledMenuItemContainerStyle,
+    css`
+      background-color: ${palette.black};
+    `,
+  ),
+};
+
+export const disabledTextStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    color: ${palette.gray.dark1};
+    font-weight: 400;
+  `,
+  [Theme.Dark]: css`
+    color: ${palette.gray.light1};
+    font-weight: 400;
+  `,
+};
 
 /**
  * Focused styles
  */
-export const focusedMenuItemContainerStyle = css`
-  &:focus {
-    text-decoration: none;
-    background-color: ${palette.blue.dark3};
-    color: ${palette.white};
+export const focusedMenuItemContainerStyle: Record<Theme, string> = {
+  [Theme.Light]: css`
+    &:focus {
+      text-decoration: none;
+      background-color: ${palette.blue.dark3};
+      color: ${palette.white};
 
-    &:before {
-      background-color: ${palette.blue.light1};
+      &:before {
+        background-color: ${palette.blue.light1};
+      }
     }
-  }
 
-  &::-moz-focus-inner {
-    border: 0;
-  }
-`;
+    &::-moz-focus-inner {
+      border: 0;
+    }
+  `,
+  [Theme.Dark]: css`
+    &:focus {
+      text-decoration: none;
+      background-color: ${palette.blue.light2};
+      color: ${palette.white};
 
-export const getFocusedStyles = (selector: string) => {
+      &:before {
+        background-color: ${palette.blue.base};
+      }
+    }
+
+    &::-moz-focus-inner {
+      border: 0;
+    }
+  `,
+};
+
+export const getFocusedStyles = (selector: string, theme: Theme) => {
   return {
     textStyle: css`
       ${selector}:focus & {
-        color: ${palette.white};
+        color: ${theme === Theme.Light ? palette.white : palette.blue.dark1};
       }
     `,
     descriptionStyle: css`
       ${selector}:focus & {
-        color: ${palette.blue.light3};
+        color: ${theme === Theme.Light
+          ? palette.blue.light3
+          : palette.blue.dark1};
       }
     `,
     iconStyle: css`
       ${selector}:focus > & {
-        color: ${palette.blue.light3};
+        color: ${theme === Theme.Light
+          ? palette.blue.light3
+          : palette.blue.base};
       }
     `,
   };
@@ -218,5 +346,21 @@ export const getFocusedStyles = (selector: string) => {
 
 export const linkStyle = css`
   text-decoration: none;
-  color: inherit;
 `;
+
+export const focusedSubMenuItemBorderStyles: Record<Theme, string> = {
+  [Theme.Light]: css`
+    &:focus {
+      &::after {
+        background-color: ${palette.blue.dark3};
+      }
+    }
+  `,
+  [Theme.Dark]: css`
+    &:focus {
+      &::after {
+        background-color: ${palette.blue.light2};
+      }
+    }
+  `,
+};